HDDS-14005. Removed duplicated FooPlugin/BarPlugin test plugin implementations
Moved FooPlugin to a simple inline implementation and made BarPlugin
(required for a multi-plugin test) simply extend from FooPlugin
