--- hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/eventlistener/BarPlugin.java
@@ -1,57 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.hadoop.ozone.om.eventlistener;
-
-import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-
-/**
- * Test utility for creating a dummy plugin.
- */
-public class BarPlugin implements OMEventListener {
-
-  private boolean initialized = false;
-  private boolean started = false;
-  private boolean shutdown = false;
-
-  @Override
-  public void initialize(OzoneConfiguration conf, OMEventListenerPluginContext pluginContext) {
-    initialized = true;
-  }
-
-  @Override
-  public void start() {
-    started = true;
-  }
-
-  @Override
-  public void shutdown() {
-    shutdown = true;
-  }
-
-  public boolean isInitialized() {
-    return initialized;
-  }
-
-  public boolean isStarted() {
-    return started;
-  }
-
-  public boolean isShutdown() {
-    return shutdown;
-  }
-}
